+++ apps/api/src/question/question.module.ts
@@ -3,9 +3,10 @@ import { QuestionController } from './question.controller';
 import { QuestionService } from './question.service';
 import { PrismaModule } from '../prisma/prisma.module';
 import { ActionHistoryModule } from '../action-history/action-history.module';
+import { ClassModule } from '../class/class.module';
 
 @Module({
-  imports: [PrismaModule, ActionHistoryModule],
+  imports: [PrismaModule, ActionHistoryModule, ClassModule],
   controllers: [QuestionController],
   providers: [QuestionService],
   exports: [QuestionService],