+++ apps/api/src/topic/topic.module.ts
@@ -1,6 +1,7 @@
 import { Module } from '@nestjs/common';
 import { PrismaModule } from 'src/prisma/prisma.module';
 import { ActionHistoryModule } from 'src/action-history/action-history.module';
+import { ClassModule } from 'src/class/class.module';
 import {
   CourseChapterController,
   CourseTopicController,
@@ -13,7 +14,7 @@ import {
 import { TopicService } from './topic.service';
 
 @Module({
-  imports: [PrismaModule, ActionHistoryModule],
+  imports: [PrismaModule, ActionHistoryModule, ClassModule],
   controllers: [
     CourseChapterController,
     CourseTopicController,